tests: add missing unsigned long casts to pthread_self()

+++ tests/test_urcu_hash_rw.c
@@ -111,7 +111,7 @@ void *test_hash_rw_thr_reader(void *_count)
 			"reader", (unsigned long) pthread_self(),
 			(unsigned long) gettid());
 	printf_verbose("readid : %lx, lookupfail %lu, lookupok %lu\n",
-			pthread_self(), URCU_TLS(lookup_fail),
+			(unsigned long) pthread_self(), URCU_TLS(lookup_fail),
 			URCU_TLS(lookup_ok));
 	return ((void*)1);
