+++ tests/test_urcu_wfs.c
@@ -213,7 +213,7 @@ fail:
 	printf_verbose("enqueuer thread_end, thread id : %lx, tid %lu, "
 		       "enqueues %llu successful_enqueues %llu, "
 		       "empty_dest_enqueues %llu\n",
-		       pthread_self(),
+			(unsigned long) pthread_self(),
 			(unsigned long) gettid(),
 		       URCU_TLS(nr_enqueues),
 		       URCU_TLS(nr_successful_enqueues),
@@ -308,7 +308,7 @@ static void *thr_dequeuer(void *_count)
 	printf_verbose("dequeuer thread_end, thread id : %lx, tid %lu, "
 		       "dequeues %llu, successful_dequeues %llu "
 		       "pop_all %llu pop_last %llu\n",
-		       pthread_self(),
+			(unsigned long) pthread_self(),
 			(unsigned long) gettid(),
 		       URCU_TLS(nr_dequeues), URCU_TLS(nr_successful_dequeues),
 		       URCU_TLS(nr_pop_all),
